Revert "fix: merge branch Self hosted/locket dio (#10)"
This reverts commit 8799d2f542c16641bc6d40d5254be1ca8769b142.
