+++ apps/self-hosted/web/.env.example
@@ -3,7 +3,7 @@
 # ==============================
 
 VITE_BASE_API_URL=http://localhost:5001
-VITE_BETA_API_URL=http://localhost:5001
+VITE_BETA_API_URL=https://api-beta.locket-dio.com
 VITE_AUTH_API_URL=http://localhost:5001
 
 # ==============================
@@ -23,11 +23,8 @@ VITE_PAYMENT_API_URL=http://localhost:5001
 # ==============================
 
 VITE_STORAGE_API_URL=http://localhost:5003
-<<<<<<< HEAD
-=======
 VITE_MEDIA_API_URL=http://localhost:5001
 VITE_CDN_URL=http://localhost:5001
->>>>>>> main
 
 # ==============================
 # 🔹 External Locket APIs
