+++ apps/self-hosted/web/src/layouts/mainLayout.jsx
@@ -4,6 +4,7 @@ import Header from "@/components/Header";
 import Footer from "@/components/Footer";
 import Sidebar from "@/components/Sidebar";
 
+const FloatingActions = lazy(() => import("@/components/ui/FloatingWidget"));
 const CropImageStudio = lazy(() =>
   import("@/components/common/CropImageStudio")
 );
@@ -30,6 +31,7 @@ const DefaultLayout = ({ children }) => {
 
       <Suspense fallback={null}>
         <CropImageStudio />
+        {!shouldHideFloating && <FloatingActions />}
       </Suspense>
 
       <Sidebar />
