+++ apps/self-hosted/web/src/services/LocketDioServices/RequestServices.js
@@ -139,6 +139,63 @@ export const getOutgoingRequestFriend = async (
   }
 };
 
+export const rejectMultipleFriendRequests = async (
+  uidList,
+  direction = "incoming",
+  batchSize = 50,
+) => {
+  try {
+    const batches = chunkArray(uidList, batchSize);
+
+    let successCount = 0;
+    let successUidList = [];
+
+    for (const batch of batches) {
+      const promises = batch.map((uid) => {
+        const body = { data: { user_uid: uid, direction } };
+        return instanceLocketV2
+          .post("deleteFriendRequest", body)
+          .then(() => uid); // nếu thành công thì trả lại uid
+      });
+
+      const responses = await Promise.allSettled(promises);
+
+      responses.forEach((r) => {
+        if (r.status === "fulfilled") {
+          successCount++;
+          successUidList.push(r.value);
+        }
+      });
+
+      // tránh spam server
+      await new Promise((r) => setTimeout(r, 500));
+    }
+
+    return { successCount, successUidList, total: uidList.length };
+  } catch (error) {
+    console.error("❌ Lỗi khi xoá lời mời:", error.message);
+    return { successCount: 0, successUidList: [], total: uidList.length };
+  }
+};
+
+export const rejectFriendRequests = async (uid, direction = "outgoing") => {
+  try {
+    const body = {
+      data: {
+        user_uid: uid,
+        direction: direction,
+      },
+    };
+
+    const response = await instanceLocketV2.post("deleteFriendRequest", body);
+
+    return response; // giả sử response trả về dữ liệu thành công
+  } catch (error) {
+    console.error("Lỗi khi xoá lời mời:", error.message);
+    return [];
+  }
+};
+
 export const SendRequestToFriend = async (uid) => {
   try {
     const response = await api.post(
@@ -151,6 +208,29 @@ export const SendRequestToFriend = async (uid) => {
   } catch (error) {
     console.error("❌ Lỗi khi tìm bạn:", error.response?.data || error.message);
     throw error;
+  }
+};
+
+export const AcceptRequestToFriend = async (uid) => {
+  try {
+    const body = { data: { user_uid: uid } };
+
+    const response = await instanceLocketV2.post("acceptFriendRequest", body);
+
+    const acceptedUid = response?.data?.result?.data?.user_uid || uid;
+    if (!acceptedUid) throw new Error("Không nhận được UID hợp lệ từ server");
+    // ✅ Lấy chi tiết user từ API
+    const newFriend = await fetchUserV2(acceptedUid);
+    // ✅ Chuẩn hoá dữ liệu friend
+    const normalized = normalizeFriendDataV2(newFriend);
+    // ✅ Trả về kết quả đồng nhất
+    return normalized;
+  } catch (error) {
+    console.error(
+      "❌ Lỗi khi chấp nhận lời mời:",
+      error.response?.data || error.message,
+    );
+    return null; // fallback an toàn
   }
 };
