--- apps/self-hosted/web/src/services/LocketServices/index.js
@@ -1,6 +1,0 @@
-export * from "./auth.services";
-export * from "./moment.services.js";
-export * from "./chat.services";
-export * from "./rollcall.services";
-export * from "./fetch.services"
-export * from "./friend.services";
