+++ apps/self-hosted/web/src/services/index.js
@@ -1,3 +1,2 @@
 export * from "./ExtensionsServices";
 export * from "./LocketDioServices";
-export * from "./LocketServices";
